% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{print.eth_signature}
\alias{print.eth_signature}
\title{Print an \code{eth_signature}}
\usage{
\method{print}{eth_signature}(x, ...)
}
\arguments{
\item{x}{(eth_signature) the signature to print.}

\item{...}{Unused; for S3 compatibility.}
}
\value{
Invisibly \code{x}.
}
\description{
Print an \code{eth_signature}
}
